Fix(InputLotto): String형으로부터 int형으로 바꾸어서 반환하던 toInts 메서드를 String형으로부터 long형으로 반환하도록 수정

+++ src/main/java/lotto/Lotto.java
@@ -20,7 +20,7 @@ public class Lotto {
         }
     }
 
-    public int lottoCount(int amount) {
+    public long lottoCount(long amount) {
         exceptionHandling.notDivideException(amount); // 예외 검사
         return amount / 1000;
     }
